+++ Pizza-Android/app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".ui.IntroActivity"
+            android:name=".view.IntroActivity"
             android:exported="true">
             <meta-data
                 android:name="android.app.lib_name"
