+++ Pizza-Android/app/src/main/java/android/ui/pizza/view/food/FoodActivity.java
@@ -1,4 +1,4 @@
-package android.ui.pizza.ui.food;
+package android.ui.pizza.view.food;
 
 
 import androidx.appcompat.app.AppCompatActivity;
