+++ Pizza-Android/app/src/main/java/android/ui/pizza/view/home/HomeFragment.java
@@ -1,10 +1,9 @@
-package android.ui.pizza.ui.home;
+package android.ui.pizza.view.home;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
