+++ Pizza-Android/app/src/main/java/android/ui/pizza/view/home/adapter/RatingAdapter.java
@@ -1,6 +1,6 @@
-package android.ui.pizza.ui.home.adapter;
+package android.ui.pizza.view.home.adapter;
 
-import android.ui.pizza.ui.home.viewholder.RatingViewHolder;
+import android.ui.pizza.view.home.viewholder.RatingViewHolder;
 import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
