+++ Pizza-Android/app/src/main/java/android/ui/pizza/view/shopping/ShoppingViewModel.java
@@ -1,16 +1,16 @@
-package android.ui.pizza.ui.notifications;
+package android.ui.pizza.view.shopping;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class ShoppingViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public ShoppingViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is Shopping fragment");
     }
 
     public LiveData<String> getText() {
