+++ Pizza-Android/app/src/main/res/layout/activity_intro.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.IntroActivity">
+    tools:context=".view.IntroActivity">
 
     <FrameLayout
         android:id="@+id/fl_try"
