+++ Pizza-Android/app/src/main/res/layout/fragment_shopping.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.shopping.ShoppingFragment">
+    tools:context=".view.shopping.ShoppingFragment">
 
     <TextView
         android:id="@+id/text_dashboard"
